upload changes

+++ src/Components/Posts/PostDetails.jsx
@@ -16,40 +16,81 @@ import { BsThreeDotsVertical } from "react-icons/bs";
 import { IoSettings } from "react-icons/io5";
 import { FaTrashCan } from "react-icons/fa6";
 import { FaArchive } from "react-icons/fa";
+import Swal from 'sweetalert2';
 
 
 const PostDetails = () => {
   const likeCommentStyle = "md:text-xl w-full active:scale-95 transition-all px-4 py-2 rounded-md  hover:bg-zinc-300 active:bg-zinc-300 cursor-pointer flex items-center gap-2"
-
+  const [likesCount, setlikesCount] = useState(0)
   const post = useLoaderData()
   const { userData, user, postsData, setPostsData } = useContext(AuthContext)
   const { name, profilephotourl } = userData;
   const navigate = useNavigate()
 
+
+
   const deletePost = () => {
-    // fetch(`http://localhost:3000/post/delete/${post._id}`, {
-    fetch(`https://mini-social-app-backend.vercel.app/post/delete/${post._id}`, {
-      method: 'DELETE',
-    })
-      .then(res => res.json())
-      .then(data => {
 
-        const remaining = postsData.filter(post => post._id != data._id)
-        setPostsData(remaining)
+    const swalWithTailwind = Swal.mixin({
+      customClass: {
+        confirmButton: "bg-green-600 hover:bg-green-700 ml-2 cursor-pointer text-white font-bold py-2 px-4 rounded mr-2",
+        cancelButton: "bg-red-600 hover:bg-red-700 mr-2 cursor-pointer  text-white font-bold py-2 px-4 rounded"
+      },
+      buttonsStyling: false
+    });
 
-        navigate(`/profile/${user.email}`)
-        console.log("Post Delete Successfully", data)
-      })
+    swalWithTailwind.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Yes, delete it!",
+      cancelButtonText: "No, cancel!",
+      reverseButtons: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+
+
+        fetch(`https://mini-social-app-backend.vercel.app/post/delete/${post._id}`, {
+          method: 'DELETE',
+        })
+          .then(res => res.json())
+          .then(data => {
+
+            setShowEdit(!showEdit)
+            if (data.deletedCount > 0) {
+              swalWithTailwind.fire({
+                title: "Deleted!",
+                text: "Your file has been deleted.",
+                icon: "success"
+              });
+              navigate('/profile')
+              const remaining = postsData.filter(posts => posts._id != post._id)
+              setPostsData(remaining)
+            }
+
+          })
+
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        swalWithTailwind.fire({
+          title: "Cancelled",
+          text: "Your imaginary file is safe :)",
+          icon: "error"
+        });
+      }
+    });
+
+
   }
 
   const [like, setlike] = useState(0)
   const [showEdit, setShowEdit] = useState(1)
 
   return (
-    <div className='h-screen md:px-10 md:pt-5 pt-20 pb-3 scroll-auto md:mx-auto mx-3  space-y-5  flex justify-center items-center md:flex-row flex-col md:gap-10'>
+    <div className='flex md:flex-row flex-col md:mt-5 mt-20 md:mx-0 mx-3 gap-5 md:ml-5 '>
 
       {/* post container */}
-      <div className="md:w-full shadow-xl border border-zinc-300 rounded-2xl md:rounded-3xl bg-white">
+      <div className="md:w-full  shadow-xl border border-zinc-300 rounded-2xl md:rounded-3xl bg-white">
 
         {/* post author details  */}
         <div className="md:px-5 md:py-3 p-3 flex justify-between items-center">
@@ -72,7 +113,6 @@ const PostDetails = () => {
               <BsThreeDotsVertical className="cursor-pointer active:scale-95 hover:bg-zinc-300 active:bg-zinc-300 text-4xl text-zinc-500 hover:text-black  rounded-full transition-all p-2" />
             </button>
 
-
             <div onClick={() => { setShowEdit(!showEdit) }} className={`absolute right-9 bg-white w-50 border border-zinc-300 shadow-2xl p-3  rounded-md space-y-1 transition-all duration-500 ${showEdit ? '-z-10 opacity-0' : ' opacity-100 z-10'}`} >
               <button className={likeCommentStyle}>
                 <h1 className='flex justify-center items-center gap-2  text-sm '> {<FaBookmark />} Save post</h1>
@@ -91,11 +131,7 @@ const PostDetails = () => {
                 <h1 className='flex justify-center items-center gap-2 text-sm '> {<FaTrashCan />} Move to trash</h1>
               </button>
             </div>
-
-
           </div>
-
-
         </div>
 
         <hr className="text-zinc-300" />
@@ -106,7 +142,7 @@ const PostDetails = () => {
           <div className='rounded-lg md:w-full space-y-3 '>
             <h1 className="space-x-2 md:text-md flex flex-wrap">{post?.postContent}</h1>
             <div className="rounded-lg overflow-hidden">
-              <img className="hover:scale-105 w-full h-full active:scale-150 active:cursor-zoom-in duration-500 transition-all" src={post.postImageUrl} alt="" />
+              <img className="hover:scale-105 w-full md:h-[600px]  h-[280px] active:scale-150 active:cursor-zoom-in duration-500 transition-all" src={post.postImageUrl} alt="" />
             </div>
           </div>
 
@@ -115,11 +151,11 @@ const PostDetails = () => {
             {/* buttons  */}
             <div className="flex items-center md:gap-8 gap-6">
 
-              <button onClick={() => { setlike(!like) }} className={likeCommentStyle}>
+              <button onClick={() => { setlike(!like), setlikesCount(likesCount + 1) }} className={likeCommentStyle}>
                 <div className="text-2xl  cursor-pointer active:scale-95 transition-all active:text-black">
                   {like ? <BiSolidLike /> : < BiLike />}
                 </div>
-                <span className="flex items-center gap-2">12 <span className="hidden md:flex">Likes</span></span>
+                <span className="flex items-center gap-2">{likesCount} <span className="hidden md:flex">Likes</span></span>
               </button>
 
 
@@ -173,7 +209,7 @@ const PostDetails = () => {
 
       </div>
 
-      <div className='md:w-1/3 w-full flex justify-center items-center border border-zinc-300 h-full shadow-2xl rounded-md'>
+      <div className='md:w-1/3 w-full h-full flex justify-center items-center border border-zinc-300 shadow-2xl rounded-md'>
         <h1>Comment Loading</h1>
       </div>
     </div>
